// src/common/enum/notification-type.enum.ts
export enum NotificationType {
  APPOINTMENT_CREATED = 'APPOINTMENT_CREATED',
  APPOINTMENT_UPDATED = 'APPOINTMENT_UPDATED',
  APPOINTMENT_DELETED = 'APPOINTMENT_DELETED',
  APPOINTMENT_PAID = 'APPOINTMENT_PAID',
  APPOINTMENT_STARTED = 'APPOINTMENT_STARTED',
  APPOINTMENT_COMPLETED = 'APPOINTMENT_COMPLETED',
  APPOINTMENT_CANCELLED = 'APPOINTMENT_CANCELLED',
  APPOINTMENT_RESCHEDULED = 'APPOINTMENT_RESCHEDULED',
  APPOINTMENT_NO_SHOW = 'APPOINTMENT_NO_SHOW',
  APPOINTMENT_LATE = 'APPOINTMENT_LATE',
  APPOINTMENT_REMINDER_6H = 'APPOINTMENT_REMINDER_6H',
  APPOINTMENT_REMINDER_1H = 'APPOINTMENT_REMINDER_1H',
  PAYMENT_SUCCESS = 'PAYMENT_SUCCESS',
  DOCTOR_VERIFIED = 'DOCTOR_VERIFIED',
  DOCTOR_REJECTED = 'DOCTOR_REJECTED',
  DOCTOR_SUSPENDED = 'DOCTOR_SUSPENDED',
  DOCTOR_UNSUSPENDED = 'DOCTOR_UNSUSPENDED',
}